backend/pyproject.toml
[tool.poetry]
name = "word2vec"
version = "0.1.0"
description = ""
authors = ["Ez <ezioxxk@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
gensim = "^4.1.2"


uvicorn = "^0.17.0"

requests = "^2.27.1"
pytest = "^7.1.0"
motor = "^3.0.0"
scipy = "^1.8.1"
python-multipart = "^0.0.5"
SQLAlchemy = "^1.4.37"
psycopg2-binary = "^2.9.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^4.3.4"


[tool.poetry.dev-dependencies]
fastapi = "^0.71.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
